Separated WorkPathTranslate code to separate class
